wager table player migration fix

+++ database/migrations/2025_09_26_000000_create_wager_bets_table_simple.php
@@ -15,10 +15,10 @@ return new class extends Migration
             $table->foreignId('wager_id')->constrained()->onDelete('cascade');
             $table->foreignId('wager_choice_id')->constrained()->onDelete('cascade');
             $table->foreignId('wager_player_id')->constrained()->onDelete('cascade');
-            $table->decimal('bet_amount', 8, 2);
-            $table->decimal('amount', 8, 2);
+            $table->integer('bet_amount');
+            $table->integer('amount');
             $table->string('status')->default('pending');
-            $table->decimal('actual_payout', 8, 2)->nullable();
+            $table->integer('actual_payout')->nullable();
             $table->timestamps();
         });
     }
